skeleton now displays TEXT text section

+++ libft/src/ft_flush_buf_overflow.c
@@ -6,7 +6,7 @@
 /*   By: lgarczyn <lgarczyn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/05/08 21:34:00 by lgarczyn          #+#    #+#             */
-/*   Updated: 2015/05/08 21:34:07 by lgarczyn         ###   ########.fr       */
+/*   Updated: 2018/01/25 20:50:08 by lgarczyn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@
 void				ft_flush_buf_overflow(void)
 {
 	static t_buffer	*buffer;
-	size_t			ret;
+	ssize_t			ret;
 
 	if (!buffer)
 		buffer = ft_buf(NULL, 0, 0);
 	ret = write(buffer->fd, buffer->buffer, buffer->pos);
-	if (ret == (size_t)-1)
+	if (ret == -1)
 		ft_perror(NULL);
 	buffer->pos = 0;
 }
